Fix front_end version number in preparation for publishing.
Analyzer needs to depend on a published version of front_end in order to
pass buildbot tests (it may be an alpha version).  This CL sets the
version to "0.1.0-alpha.0" (I will publish this version once the CL
lands).  Note that the previous version of front_end was "0.1.0", so we
appear to be moving the version backwards.  This is ok because version
"0.1.0" was never published.

Note that during this transitional period where code is being moved from
analyzer to front_end, we have circular imports between analyzer and
front_end, so from now until the transitional period ends, we will need
to publish front_end whenever we publish analyzer.

Update copyright notices in front_end package

R=brianwilkerson@google.com, scheglov@google.com

Review URL: https://codereview.chromium.org/2485993003 .