Fix HTypeConversion.checkedInput
For function types, HTypeConversion.checkedInput would return the call
to the method on the reified type object.  Now the construction of the
call is deferred to codegen.dart, so the input is available directly.

The optimizer also detects checks (type assertions and 'as' casts) on
`null` values and removes the useless test.

BUG=
R=efortuna@google.com

Review URL: https://codereview.chromium.org/2497313003 .