VM: [Kernel] Cherry-pick from dart-lang/kernel_sdk
This is part of the following CL:

Roll Kernel to fix an async bug.

There was a bug in the translation of await in finally when there
was a return in the try block.

The new async/async* translation revealed a bug in the scoping of
labels.  Labels from enclosing functions were visible in nested
functions.  That bug is fixed here.

Review URL: https://chromereviews.googleplex.com/536237014 .

R=vegorov@google.com

Review URL: https://codereview.chromium.org/2515323002 .