Deoptimize Search - request a new resolved unit for local search.
It seems to me that it would be easier to integrate if we keep the
same search technique - search for Element references. I think any
saving can be only for local elements search. If it turns out to
be expensive, we could add priority results caching in the driver.

R=brianwilkerson@google.com
BUG=

Review URL: https://codereview.chromium.org/2521913004 .