Avoid random stacktraces being printed from tools/test.py during test runs
Previously the testing script was just printing interleaved stack traces with
all the other output. Which means it's not associatable with a single test and
is highly confusing!

This CL associates the stack trace information of tests which time out to the
particular test which timed out.

R=whesse@google.com

Review URL: https://codereview.chromium.org/2524733004 .