Remove the ability in dart2js to turn off syntax-only generic methods.
This CL removes all the checks for 'enableGenericMethodSyntax' in the
implementation, which takes away the ability to turn this feature off
and simplifies the code.

As a result, `ParserOptions` is now empty; we may want to remove it,
but this CL does not do that.

R=sigmund@google.com

Review URL: https://codereview.chromium.org/2521073003 .