Fix some puzzling code in object.cc.
Jens was looking through object.cc and he spotted:

    if (Original()) ...

where it seems like Original returns a RawField* which is never 0, so
it is always true.  It looks like it was intended to be a call to
IsOriginal().

(See https://codereview.chromium.org/1722733002/diff/120001/runtime/vm/object.cc)

BUG=

Review URL: https://codereview.chromium.org/2525633003 .