Filter out AnalysisErrors with null 'errorCode'.
We have an error report:

My guess is that `errorCodeByUniqueName` returned `null`.

It might be either a bug in `errorCodeByUniqueName`, or a shortcoming
or our current data version mechanism (which we know about - it should
be a hash of sources or something like this).

Either way, a missing error is better than a crash.

R=paulberry@google.com, brianwilkerson@google.com
BUG=

Review URL: https://codereview.chromium.org/2527423002 .