Search reference to a private element only in its library files.
A bit builky, not quite sure if it worth it.

Potentally useful if there are many libraries declaring private methods
with the same name, invoked outside of the declaring class (where the
method is not external, so is not recorded as a reference). But it is
an open question whether this often happens in practice.

R=paulberry@google.com, brianwilkerson@google.com
BUG=

Review URL: https://codereview.chromium.org/2544993005 .