Report CompileTimeErrorCode.URI_DOES_NOT_EXIST when the file does not exist.
When we don't call SummaryDataStore.addLinkedLibrary() for the corresponding
file, SummaryResynthesizer.getLibrary() creates a synthetic LibraryElement.
When a LibraryElement is synthetic, ImportsVerifier does not add it,
so HintCode.UNUSED_IMPORT is not generated.

And because there is no linked summary for the source,
InputPackagesResultProvider.hasResultsForSource() returns `false`, so
we attempt to read the source, get modification time `-1` and report
CompileTimeErrorCode.URI_DOES_NOT_EXIST in the VerifyUnitTask.
validateReferencedSource().

R=brianwilkerson@google.com, paulberry@google.com
BUG=

Review URL: https://codereview.chromium.org/2549223002 .