Implement 'Import Library' Quick Fix for the new analysis driver.
It turned out to be easier than I thought is would be.

The potential problem was to find correspondence between AnalysisResult
and the AnalysisDriver instance to request top-level declarations in.
But if we just include the AnalysisDriver into AnalysisResult, this
solves the problem.

This creates another problem though - now we need to be careful to
not cache an instance of AnalysisResult for a long time.

R=brianwilkerson@google.com, paulberry@google.com
BUG=

Review URL: https://codereview.chromium.org/2549143007 .