Transition to analyzing/idle only when there is/not files to analyze.
So that the analysis status is now telling the user that there might be
new errors in not yet analyzed files, not just that there is some
technical work to do in background.

This makes Element type hierarchy building about 25% faster because we avoid sending server notifications several hundred times.

R=brianwilkerson@google.com, paulberry@google.com
BUG=

Review-Url: https://codereview.chromium.org/2559333002 .