When a file is removed from the driver, reschedule full analysis.
Removing files happens rarely, so it should be OK to keep it simple
and just reset the file state and let analysis to happen from scratch.

It look in profiler that the most expensive operation is recomputing
content hash (16%) and unlinked bundle key hash (14%). Reading content
itself is somewhat expensive (11%) on SSD. Might be higher on NFS.

R=brianwilkerson@google.com, paulberry@google.com
BUG=

Review-Url: https://codereview.chromium.org/2566003002 .