Do not generate a duplicate ExternalName class
In https://github.com/dart-lang/sdk/commit/f7b6f0cf1c188bbc375de2a37f9a4129cb845428
some of the VM's libraries were modified to work the the analyzer's
patching support.  This requires the class ExternalName, which used as
an annotation, to be present in the SDK.

Before: it was injected by dart code as part of a build step.  Now, it
just appears as code in the SDK and we should not inject a (duplicate)
definition.

BUG=
R=kustermann@google.com

Review-Url: https://codereview.chromium.org/2580783002 .