Start re-trying dartk compilations if they crash
Every crash of a dartk invocation will cause 1 more re-try. This is independent
of whether the crash was caused by the entire batch runner or just that single
compilation caused an exception.

The unfortunate side-effect of this is that ./tools/test.py will now print
warnings every time it retries the command.

R=vegorov@google.com

Review-Url: https://codereview.chromium.org/2576693002 .