dart2js kernel: do less enqueuing.
Convert to kernel only classes and elements that have been resolved. For
anything else, pretend the member doesn't exist.

One exception: fields are added because an unused field might still be stored in
an initializer. This appears to only occur with constants in metadata. Creating the constant in code will not hit this issue.

This makes the compiler do work for the same set of elements that were included during resolution. Running all language tests with --dart2js-batch is now 2.5 min (down from 8 min)

BUG=
R=efortuna@google.com

Review-Url: https://codereview.chromium.org/2604463002 .