Gardening: mark regexp pcre_test slow/timeout in precompiled code
This test seems to be slow in any configuration. There are old bugs,
e.g. 21593 and 22008 that mention the slowness and flakiness, but there
are no takers to either fix the regexp code or make the test smaller.

Note: in precompiled mode, the regular expression code is interpreted, and thus actually slower than in the JIT version of the test.

BUG=
R=rmacnak@google.com

Review-Url: https://codereview.chromium.org/2624733005 .