Fix physical_file_system_test.dart on Windows.
With the simplification to PhysicalFileSystem introduced in
fef4569eb5a80c72b1f7b89dfac96cd78f2b19cf, the URI "/foo.dart" is now
uniformly treated as equivalent to "file:///foo.dart", even on
Windows.  (At the time the file is read, the file will be located on
the current drive).  This is consistent with what we do for relative
paths.

TBR=sigmund@google.com

Review-Url: https://codereview.chromium.org/2634543003 .