Relax limitations on analyzing non-Dart files.
As an alternative to fixing NPE in search, we could just allow
AnalysisDriver methods return results for any, not just Dart files.

The only really important check if in addFile() so that we don't
attempt to analyze non-Dart files just because they are in the folder.
And even this check could be moved to Analysis Server, but for now it
seems easier to keep it in AnalysisDriver.

R=brianwilkerson@google.com
BUG=

Review-Url: https://codereview.chromium.org/2653453002 .