Print stack traces of tests timing out on Windows
Uses cdb.exe (part of Windows SDK, comes with depot_tools) to dump stack traces of all threads when a test times out similar
to what we already do on Linux and MacOS.

Note that cdb.exe comes in two version (32- and 64-bit). For proper stacktraces the correct version must be in the PATH.

Fixes #28242.

R=kustermann@google.com

Review-Url: https://codereview.chromium.org/2670763002 .