Connect fasta's scanner to the analyzer-derived scanner tests in front_end.
Failing tests are currently marked with "@failingTest" to prevent
breaking buildbots.

I've marked the issues I'm aware of with "TODO(paulberry,ahe)".  Peter
and I need to triage these issues--some of them may not be necessary
to fix due to the fact that fasta replaces the parser as well.  We
will follow up with CLs that either fix the issues or remove the TODO
comments as appropriate.

R=ahe@google.com

Review-Url: https://codereview.chromium.org/2676083002 .